Model: Remove ModelBase class
The ModelBase class is unnecessary now that simulation is templated on
the agent type. Therefore, it has been removed.

Further, `model.n_iterations()` is now a getter function for
`model._n_iterations` and there is a new `mode.initialize_iterations()`
function that contains behaviour that needs to be exectued before a
`simulation.run()` is started.