ActivityDrivenModel: Include reluctance
The ActivityDrivenModel driven model has been extended with a concept,
that we call reluctance. This quantifies how receptive to others'
opinions an agent is.
Reluctance is an per agent quantity and therefore the IO had to be
changed slightly.
The bots can be thought of as agents with infinite reluctance.

TODO: generation of distributions for reluctance, possibly correlated
with activities.

Co-authored-by: Amrita Goswami <amrita16thaug646@gmail.com>